% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompts.R
\name{default_eval_guidelines}
\alias{default_eval_guidelines}
\title{Default guidelines for translation.}
\usage{
default_eval_guidelines()
}
\value{
A character string containing default evaluation guidelines.
}
\description{
Evaluation guidelines that are added to the LLM prompt
}
