% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompts.R
\name{default_eval_task}
\alias{default_eval_task}
\title{Default guidelines for translation.}
\usage{
default_eval_task()
}
\value{
A character string containing default evaluation guidelines.
}
\description{
Evaluation task that is added to the LLM prompt
}
