% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompts.R
\name{default_guidelines}
\alias{default_guidelines}
\title{Default guidelines for translation.}
\usage{
default_guidelines()
}
\value{
A character string containing default translation guidelines.
}
\description{
Translation guidelines that are added to the LLM prompt
}
