% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_to_json.R
\name{dt_to_json}
\alias{dt_to_json}
\title{Convert a data.table to JSON format}
\usage{
dt_to_json(dt)
}
\arguments{
\item{dt}{A data.table object containing 'Source' and 'Target' columns}
}
\value{
A JSON string with the converted data
}
\description{
This function converts a data.table with 'Source' and 'Target' columns
into a specific JSON format with 'source_item' and 'target_item' keys.
}
\examples{
dt <- data.table(
  Source = c("How satisfied are you with your current job?"),
  Target = c("Quel est votre degré de satisfaction à l'égard de votre emploi actuel ?")
)
json_result <- dt_to_json(dt)

}
