% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_translate.R
\name{llm_translate}
\alias{llm_translate}
\title{llm translate}
\usage{
llm_translate(
  source_file,
  example_file = "Sample surveys translation.xlsx",
  example_type = "Items"
)
}
\arguments{
\item{source_file}{A single string, path to a *.txt file with items that need to be translated. Each line is one item.}

\item{example_file}{A single string. Optional.}

\item{example_type}{A single string. Optional.}
}
\value{
A data.table containing translated items. May raise an error if the LLM
call fails or returns invalid JSON.
}
\description{
A short description...
}
