% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompts.R
\name{prompt_back_trans}
\alias{prompt_back_trans}
\title{Prompt back translation}
\usage{
prompt_back_trans(JSON_DATA)
}
\arguments{
\item{JSON_DATA}{A single string containing JSON data with the columns
original, translation,  back_translation.}
}
\value{
A single string representing a prompt structure including the input JSON data.
}
\description{
A short description...
}
