% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompts.R
\name{prompt_evaluation}
\alias{prompt_evaluation}
\title{Prompt evaluation}
\usage{
prompt_evaluation(
  JSON_DATA,
  guidelines = default_eval_guidelines(),
  task = default_eval_task()
)
}
\arguments{
\item{JSON_DATA}{A string containing JSON data.}

\item{guidelines}{Optional. A string containing guidelines.}
}
\value{
A string representing a JSON structure, likely for a language model prompt.
}
\description{
A short description...
}
