% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_survey.R
\name{.batch_by_size}
\alias{.batch_by_size}
\title{Batch process items by size}
\usage{
.batch_by_size(items_obj, bp, chat, sleep, tmp_path, restart)
}
\arguments{
\item{items_obj}{An object containing data to process and the batch size. Must have a `data` element with a `Text` column and a `batch_size` element.}

\item{bp}{A string template for the prompt.}

\item{chat}{A chat object or argument passed to `.llm_response`.}

\item{sleep}{A single number representing the time to sleep between batches.}

\item{tmp_path}{Path to a temporary file for saving progress.}

\item{restart}{Logical. If TRUE, the function attempts to restart from a previously saved temporary file.}
}
\value{
A `data.table` containing the results of processing items in batches.
}
\description{
Processes survey items in batches based on a specified batch size, communicating with an LLM for translation.
}
