% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_survey.R
\name{.batch_by_size}
\alias{.batch_by_size}
\title{Batch by size}
\usage{
.batch_by_size(bp, items_to_translate, batch_size, chat, sleep)
}
\arguments{
\item{bp}{A string prompt template.}

\item{items_to_translate}{A data.table or similar with a column `Text`.}

\item{batch_size}{An integer to determine the number of items per batch}

\item{sleep}{A numeric value specifying the number of seconds to pause between batches.}

\item{An}{`ellmer` chat object}
}
\value{
A data.table containing results from batch processing. The function prints
progress messages and may pause between batches.
}
\description{
A short description...
}
\keyword{internal}
