% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_survey.R
\name{.batch_by_vars}
\alias{.batch_by_vars}
\title{Title}
\usage{
.batch_by_vars(items_obj, bp, EXAMPLES, chat, sleep)
}
\arguments{
\item{items_obj}{An object containing data and batching information.}

\item{bp}{A base prompt string.}

\item{chat}{Chat context for the LLM.}

\item{sleep}{A numeric value for sleeping between batches.}
}
\value{
A data.table containing the results for each batch.
}
\description{
A short description...
}
