% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_survey.R
\name{.batch_by_vars}
\alias{.batch_by_vars}
\title{Batch items by variables}
\usage{
.batch_by_vars(
  bp,
  items_to_translate,
  batch_vars,
  get_instr,
  source_language,
  target_language,
  guidelines,
  domain,
  EXAMPLES,
  chat,
  sleep
)
}
\arguments{
\item{bp}{A base prompt.}

\item{items_to_translate}{A data.table containing items to translate,
expected to have columns "Text", "Instrument", "Topic", "Instruction", and "Response",
and also columns specified by `batch_vars` used in the `by` argument of `data.table::data.table()`.}

\item{batch_vars}{A character vector with variable names that determine groups for
batch processing,}

\item{source_language}{Character. Source language used in the survey items. Default is `"English"`.}

\item{target_language}{Character. Target language for the translations. Default is `"Norwegian"`.}

\item{guidelines}{Character string or vector. Optional custom translation}

\item{domain}{Character. Domain context for the prompt. Helps tailor the tone}

\item{sleep}{A numeric value specifying the number of seconds to pause between batches.}

\item{An}{`ellmer` chat object}
}
\value{
A data.table containing the translation results (`batch_results` from `.translate_llm`),
combined with the original batch items and a `batch_idx` column.
}
\description{
A short description...
}
\keyword{internal}
