% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_chat}
\alias{.get_chat}
\title{Get chat object}
\usage{
.get_chat(model = NULL, api_key = NULL, temperature = 0.05)
}
\arguments{
\item{model}{Optional. A model name.}

\item{api_key}{Optional. An API key.}

\item{temperature}{Optional. A single numeric value for temperature.}
}
\value{
A chat object from the `ellmer` package. Will error if no API key environment
variable (GOOGLE_API_KEY, OPENAI_API_KEY, or ANTHROPIC_API_KEY) is found.
}
\description{
A short description...
}
